+++ packages/content-toggle-react/src/index.ts
@@ -1,0 +1,1 @@
+export { ContentToggle } from "./ContentToggle";