+++ packages/feedback-react/src/Feedback.test.tsx
@@ -1,0 +1,55 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+import { Feedback } from "./Feedback";
+import userEvent from "@testing-library/user-event";
+import { SimplifiedFeedback } from "./SimplifiedFeedback";
+import { axe } from "jest-axe";
+
+const mockFn = jest.fn();
+
+beforeEach(() => {
+    jest.resetAllMocks();
+});
+
+test("should call onSubmit function with feedback value", () => {
+    render(<Feedback label="feedback" onSubmit={mockFn} />);
+    userEvent.click(screen.getByTestId("feedback-1"));
+    userEvent.click(screen.getByTestId("submit-button"));
+    expect(mockFn).toBeCalledTimes(1);
+    expect(mockFn.mock.calls[0][0]).toStrictEqual({ feedbackValue: 1, message: "" });
+});
+
+test("should call onSubmit function with feedback value and message", () => {
+    render(<Feedback label="feedback" onSubmit={mockFn} />);
+    userEvent.click(screen.getByTestId("feedback-1"));
+    userEvent.type(screen.getByTestId("feedback-text"), "This is very nice");
+    userEvent.click(screen.getByTestId("submit-button"));
+    expect(mockFn).toBeCalledTimes(1);
+    expect(mockFn.mock.calls[0][0]).toStrictEqual({ feedbackValue: 1, message: "This is very nice" });
+});
+
+test("should call onSubmit function with feedback value and message with changes", () => {
+    render(<Feedback label="feedback" onSubmit={mockFn} />);
+    userEvent.click(screen.getByTestId("feedback-1"));
+    userEvent.type(screen.getByTestId("feedback-text"), "This is very nice");
+    userEvent.click(screen.getByTestId("feedback-2"));
+    userEvent.click(screen.getByTestId("submit-button"));
+    expect(mockFn).toBeCalledTimes(1);
+    expect(mockFn.mock.calls[0][0]).toStrictEqual({ feedbackValue: 2, message: "This is very nice" });
+});
+
+test("simplifeid should call onSubmit function with feedback value and message with changes", () => {
+    render(<SimplifiedFeedback description="feedback" label="feedback" onSubmit={mockFn} />);
+    userEvent.click(screen.getByLabelText("3"));
+    userEvent.type(screen.getByTestId("feedback-text"), "This is very nice");
+    userEvent.click(screen.getByTestId("submit-button"));
+    expect(mockFn).toBeCalledTimes(1);
+    expect(mockFn.mock.calls[0][0]).toStrictEqual({ feedbackValue: 3, message: "This is very nice" });
+});
+
+test.each([Feedback, SimplifiedFeedback])("Feedback should be a11y compliant", async (Component) => {
+    const { container } = render(<Component description="feedback" label="feedback" onSubmit={mockFn} />);
+    const results = await axe(container);
+
+    expect(results).toHaveNoViolations();
+});