+++ packages/toggle-switch-react/src/index.ts
@@ -1,1 +1,2 @@
 export { ToggleSwitch } from "./ToggleSwitch";
+export { ToggleSlider } from "./ToggleSlider";