+++ portal/src/components/Card/Card.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { navigate } from "gatsby";
+import { unicode } from "@fremtind/jkl-constants-util";
 import "./card.scss";
 
 interface Props {
@@ -9,7 +10,7 @@ interface Props {
 
 export const Card: React.FunctionComponent<Props> = ({ heading, link, children }) => {
     // right arrow for internal links, up/right for external links:
-    const arrow = link.startsWith("http") ? "\u2197" : "\u2192";
+    const arrow = link.startsWith("http") ? unicode.NORTH_EAST_ARROW : unicode.RIGHTWARDS_ARROW;
     const handleClick = () => navigate(link);
 
     return (